perf: Use fewer watchers
With this new strategy, top level directories can be checked against the
gitignore rules before setting up a watcher for them. This should
significantly reduce the number of watchers on git repositories, either
from ignored directories or from the .git directory.

I don't thing having a large number of files in the root directory is a
problem, but it could be.

This works, so next step is to handle changes to the root directory
itself and update watchers.